requirements.txt
flask==3.1.0
python-dotenv==1.1.0
PyMuPDF==1.25.3
python-docx==1.1.2

# ── SaaS 后端依赖 ──
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.7.1
flask-limiter==3.11.0
flask-cors==5.0.1
flask-migrate==4.1.0
marshmallow==3.25.1
APScheduler==3.11.0
